refactor: split different sentinels into separate files
