+++ vllm/v1/engine/core_client.py
@@ -53,17 +53,13 @@ from vllm.v1.engine.utils import (
     launch_core_engines,
 )
 from vllm.v1.executor import Executor
-from vllm.v1.fault_tolerance.sentinel import ClientSentinel
+from vllm.v1.fault_tolerance import ClientSentinel
 from vllm.v1.fault_tolerance.utils import (
     FaultToleranceRequest,
     FaultToleranceResult,
     FaultToleranceZmqAddresses,
 )
-from vllm.v1.serial_utils import (
-    MsgpackDecoder,
-    MsgpackEncoder,
-    bytestr,
-)
+from vllm.v1.serial_utils import MsgpackDecoder, MsgpackEncoder, bytestr
 
 logger = init_logger(__name__)
