+++ vllm/v1/fault_tolerance/__init__.py
@@ -1,0 +1,13 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM project
+from .client_sentinel import ClientSentinel
+from .engine_core_sentinel import EngineCoreSentinel
+from .sentinel import BaseSentinel
+from .worker_sentinel import WorkerSentinel
+
+__all__ = [
+    "BaseSentinel",
+    "ClientSentinel",
+    "EngineCoreSentinel",
+    "WorkerSentinel",
+]
